% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Veritabanı bağlantısını kurar ve bağlantı nesnesini döndürür}
\usage{
connect_to_database(dbPath)
}
\arguments{
\item{dbPath}{Veritabanı dosyasının yolunu belirtir.}
}
\value{
Veritabanı bağlantısı nesnesi.
}
\description{
Veritabanı bağlantısını kurar ve bağlantı nesnesini döndürür
}
